// src/models/extra.ts
export interface Like {
  id: string
  creator: any
}

export enum PoliticalType {
  LEFT = "left",
  RIGHT = "right",
  LIBERTARIAN = "libertarian",
  CENTRIST = "centrist",
  CONSERVATIVE = "conservative",
  PROGRESSIVE = "progressive",
  SOCIALIST = "socialist",
  GREEN = "green",
  NATIONALIST = "nationalist",
  FASCIST = "fascist",
  ANARCHIST = "anarvhist",
  SOCIALDEMOCRAT = "socialDemocrat",
  COMMUNIST = "communist",
  MODERATE = "moderate",
}